renaming api_functions -> handlers
