+++ explorer/api/handlers/health_check.go
@@ -1,4 +1,4 @@
-package apifunctions
+package handlers
 
 import (
 	"context"
